+++ tests/schema/models_o2o_2.py
@@ -8,5 +8,5 @@ from tortoise.models import Model
 
 class One(Model):
     tournament: fields.OneToOneRelation[Two] = fields.OneToOneField(
-        "models.Two", on_delete="WABOOM"
+        "models.Two", on_delete="WABOOM"  # type:ignore
     )