+++ CHANGELOG.rst
@@ -10,6 +10,9 @@ Changelog
 ====
 0.19.4
 ------
+Added
+^^^^^
+- Allow ForeignKeyField(on_delete=NO_ACTION) (#1393)
 Fixed
 ^^^^^
 - Fix foreign key constraint not generated on MSSQL Server. (#1400)